fix: Make ContainerRegistryScope to ExtraScope subtype
